agent-deployer/pyproject.toml
[tool.poetry]
name = "agent-deployer"
version = "0.1.0"
description = ""
authors = ["P. Taylor Goetz <taylor@sema4.ai>"]
readme = "README.md"
#packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.8"
requests = "^2.32.2"
pyyaml = "^6.0.1"
